# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "uSocial API",
        "description": "API documentation for uSocial",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/purchase-gems": {
            "post": {
                "tags": [
                    "Gems"
                ],
                "summary": "Purchase gems for the authenticated user",
                "operationId": "a185d4728bbdf0ea5f5c2a51257a6997",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paymentMethodId",
                                    "amount"
                                ],
                                "properties": {
                                    "paymentMethodId": {
                                        "description": "Stripe payment method ID",
                                        "type": "string"
                                    },
                                    "amount": {
                                        "description": "Number of gems to purchase",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment successful"
                                        },
                                        "transaction": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "transaction_type": {
                                                    "type": "string",
                                                    "example": "purchase"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "stripe_transaction_id": {
                                                    "type": "string",
                                                    "example": "ch_1Nm6x2FEXAMPLE"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Payment failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Card declined"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/create-intent": {
            "post": {
                "tags": [
                    "Gems"
                ],
                "summary": "Create a payment intent for purchasing gems",
                "description": "Generates a payment intent for the given gem amount. Requires the user to be authenticated.",
                "operationId": "5d7531259a8654a7b480401cca6e6e7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Number of gems to purchase",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client secret for the payment intent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clientSecret": {
                                            "description": "Client secret for Stripe payment intent",
                                            "type": "string",
                                            "example": "pi_1EydQl2eZvKYlo2Cxl97"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "The amount field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Something went wrong while processing the payment intent."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/finalize-purchase": {
            "post": {
                "tags": [
                    "Gems"
                ],
                "summary": "Finalize the purchase of gems",
                "description": "Completes the gem purchase by recording the transaction and updating the user's gem balance. Requires the user to be authenticated.",
                "operationId": "c21a35d27277935c3f15d49520405296",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paymentId",
                                    "amount"
                                ],
                                "properties": {
                                    "paymentId": {
                                        "description": "Stripe payment ID",
                                        "type": "string",
                                        "example": "pi_1EyjdV2eZvKYlo2Cxl98"
                                    },
                                    "amount": {
                                        "description": "Number of gems purchased",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment finalized successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "Payment finalized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "The paymentId field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}